fix ELK stack in Makefile and compose file
in Makefile, no need to use `-f` for the logging compose file, it is
included with `include:` in the main compose file. For the same reason,
no need to specify the logging .env file separately.

Using `--profile` with `down` is safe. It will put down all containers
with no profile, and containers with the mentioned profile.

GELF logging driver is needed for hte ELK stack